Create duplicate item feature Order list by id Create dialog confirm delete Create dialog confirm duplicate Create move item feature Change icon InputText when contains text

+++ app/src/main/java/br/com/joaoov/ui/ambient/AmbientListFragment.kt
@@ -15,7 +15,8 @@ import br.com.joaoov.Path.Companion.DEPARTAMENT_PATH
 import br.com.joaoov.R
 import br.com.joaoov.data.local.ambient.Ambient
 import br.com.joaoov.data.local.departament.Departament
-import br.com.joaoov.ui.component.GenericDialog
+import br.com.joaoov.ui.component.AlertDialogCustom
+import br.com.joaoov.ui.component.openMoveDialog
 import kotlinx.android.synthetic.main.fragment_ambient.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
@@ -30,9 +31,13 @@ class AmbientListFragment : Fragment(R.layout.fragment_ambient) {
         AmbientListAdapter(
             onClick = { navigateToFunctionFragment(it) },
             onEditClick = { navigateToEditFragment(it) },
-            onDuplicateClick = { viewModel.duplicate(it) },
+            onDuplicateClick = {
+                AlertDialogCustom(requireContext())
+                    .showDuplicateDialog { viewModel.duplicate(it) }
+            },
+            onMoveClick = { openMoveDialog(it) },
             onDeleteClick = {
-                GenericDialog(requireContext())
+                AlertDialogCustom(requireContext())
                     .showDeleteDialog { viewModel.delete(it) }
             }
         )
