+++ app/src/main/java/br/com/joaoov/ui/company/CompanyListFragment.kt
@@ -11,7 +11,7 @@ import br.com.joaoov.ComponentViewModel
 import br.com.joaoov.Components
 import br.com.joaoov.R
 import br.com.joaoov.data.local.company.Company
-import br.com.joaoov.ui.component.GenericDialog
+import br.com.joaoov.ui.component.AlertDialogCustom
 import kotlinx.android.synthetic.main.fragment_ambient.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
@@ -25,7 +25,7 @@ class CompanyListFragment : Fragment(R.layout.fragment_company) {
             onClick = { navigateToDepartamentFragment(it) },
             onEditClick = { navigateToEditCompanyFragment(it) },
             onDeleteClick = {
-                GenericDialog(requireContext())
+                AlertDialogCustom(requireContext())
                     .showDeleteDialog { viewModel.delete(it) }
             }
         )
