+++ app/src/main/java/br/com/joaoov/ui/departament/DepartamentListFragment.kt
@@ -15,7 +15,7 @@ import br.com.joaoov.Path.Companion.COMPANY_PATH
 import br.com.joaoov.R
 import br.com.joaoov.data.local.company.Company
 import br.com.joaoov.data.local.departament.Departament
-import br.com.joaoov.ui.component.GenericDialog
+import br.com.joaoov.ui.component.AlertDialogCustom
 import kotlinx.android.synthetic.main.fragment_ambient.*
 import org.koin.android.viewmodel.ext.android.sharedViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
@@ -30,7 +30,7 @@ class DepartamentListFragment : Fragment(R.layout.fragment_departament) {
             onClick = { navigateToAmbientFragment(it) },
             onEditClick = { navigateToEditFragment(it) },
             onDeleteClick = {
-                GenericDialog(requireContext())
+                AlertDialogCustom(requireContext())
                     .showDeleteDialog { viewModel.delete(it) }
             }
         )
